corrected for Dom's comments
