+++ croptalk/utils.py
@@ -1,13 +1,9 @@
-import pandas as pd
-from sqlalchemy import text
-
-from croptalk.create_doc_data import engine
+from langchain.chat_models import ChatOpenAI
 
 
-def query_sql_data(query):
-    # Execute the SQL query
-    with engine.connect() as connection:
-        result = connection.execute(text(query))
-        rows = result.fetchall()
-
-    return pd.DataFrame(data=rows)
+def initialize_llm(model):
+    return ChatOpenAI(
+        model=model,
+        streaming=True,
+        temperature=0,
+    )
